Handle start and end date in async_get_events
When requesting events home assistant might specify a start and end
date that the events returned should be between (for example, if a
user has an automation that calls calendar.get_events). The previous
implementation ignored these values and so would always return a full
set of events. This adds simple handling of those values to skip
returning events that fall outside the range.

The range passed in by home assistant may include time data, but that
is not handled here because we count all pickup events as taking the
whole day.